# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FitPaxPro Gym API",
        "description": "Gym module mobile authentication endpoints.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/fitpaxpro/FitPaxPro/public",
            "description": "Application Server"
        }
    ],
    "paths": {
        "/api/admin/login": {
            "post": {
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Authenticate an admin account",
                "description": "Logs in a super admin using email and password and returns a Sanctum bearer token. Super admin users are identified by user_type = 0 in the existing system.",
                "operationId": "fff602ba09ada3bd4325f5a720a3e2a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password@123"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "swagger-ui",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Admin login successful."
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|sanctumtoken"
                                                },
                                                "admin": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Super Admin"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "format": "email",
                                                            "example": "superadmin@example.com"
                                                        },
                                                        "user_type": {
                                                            "description": "Super admin user type",
                                                            "type": "integer",
                                                            "example": 0
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation or authentication error"
                    }
                }
            }
        },
        "/api/admin/me": {
            "get": {
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Get authenticated admin profile",
                "description": "Returns the currently authenticated admin user based on the provided Sanctum bearer token.",
                "operationId": "48bfb1d39ddc90f215a0f9fbbdbd76c3",
                "responses": {
                    "200": {
                        "description": "Admin profile fetched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Admin profile fetched successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "admin": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Super Admin"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "format": "email",
                                                            "example": "superadmin@example.com"
                                                        },
                                                        "user_type": {
                                                            "type": "integer",
                                                            "example": 0
                                                        },
                                                        "status": {
                                                            "type": "boolean",
                                                            "example": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden: non-admin token"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/logout": {
            "post": {
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Logout admin and revoke current token",
                "description": "Revokes the currently used Sanctum access token for the authenticated admin user.",
                "operationId": "e9936e079bbfdf1e37ae16834aed0ff0",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Admin logout successful."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden: non-admin token"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}